refactor(backend): encapsulate preparation logic for cached sandbox
